Move contract deploy related test to `test_deploy` file
